Lazy load modal, memoize tools and remove deps
